events: Move ignored_users serde code into ignored_user_list module
… and optimize it.
