+++ crates/ruma-common/CHANGELOG.md
@@ -4,6 +4,7 @@ Breaking changes:
 
 * Remove deprecated `EventType` enum
 * Remove deprecated constructors for `RoomMessageEventContent`
+* Remove `serde::vec_as_map_of_empty` from the public API
 
 # 0.10.3
